Workaround iOS 8 UITableViewController init bug
* Make `TableViewController` inherit from `UIViewController`:

On iOS 8, `UITableViewController` incorrectly implements initialization
chaining. In its implementation of `init(style:)`, it invokes
`self.init(nibName: nibBundle:)` instead of calling up to `super`. In
Objective-C this isn't a problem since initializers are always inherited. In
Swift, adding a required initializer effectively short circuits initializer
inheritance causing the default implementation of `init(nibName: nibBundle:)`
not to be inherited from `UIViewController`. The end result is that subclasses
that add a required initializer that calls up to `init(style:)` crash on iOS 8
because that implementation then tries to invoke an `init(nibName: nibBundle:)`
method on the subclass that is not implemented. In many cases, it becomes
impossible for consumer subclasses to implement that method and have their
objects remain in a consistent state since there may not be a sensible default
value for any property that is assigned with the argument provided to their
required, designated initializer.