Remove reference to functions as selections
This was causing memory leaks as the closure referenced self, without declaring it's ownership of self.
From the Swift docs: "A strong reference cycle can also occur if you assign a closure to a property of a
class instance, and the body of that closure captures the instance...[this] strong reference cycle occurs
because closures, like classes, are reference types"

By adding the [unowned self] to each selection, we remove any retain cycles.