Fix test on iOS 11
It seems that `UITableView.cellForRow(at:)` now actually returns nil if
the requested cell is off-screen. Because our test table view is never
mounted in a view hierarchy, we never get a cell back.

Instead, we’ll ask the data source directly for the correct cell. This
is probably a better test since we’re testing our behavior directly.